requirements.txt
awswrangler==2.8.0
boto3==1.17.76
pandas==1.2.4
pendulum==2.1.2
python-decouple==3.4
requests==2.25.1